Make upload file functionality

+++ database/migrations/2020_08_17_124259_create_drives_table.php
@@ -18,7 +18,7 @@ class CreateDrivesTable extends Migration
             $table->integer('user_id');
             $table->string('file_name');
             $table->string('file_type');
-            $table->integer('file_size');
+            $table->decimal('file_size');
             $table->integer('is_star');
             $table->integer('is_trash');
             $table->timestamps();
